Consolidate client dialer under common
